Fix custom region logic (#14)
Previously, this code used Rusoto to get the default region as an object then, if there was a custom endpoint turn that region into a Region::Custom with the previously acquired default region name as the custom name. However, Region::default() performs internal checks to filter out names that do not match the official AWS region names and defaults to 'us-east-1' if it's not in the list. This result in custom region names always being ignored and defaulting to 'us-east-1'. The fix changes the logic around this to first check if a custom endpoint is used and then check for custom region names - if no custom endpoint is specified, it simply uses the AWS default.